[Refactor] interface 추가 및 랜덤 추출 코드 수정
